separating oracle interogation
